node - move sqs service to aws dir
