+++ node-server/src/auth/auth.module.ts
@@ -5,7 +5,7 @@ import { UsersModule } from '../users/users.module';
 import { JwtModule } from '@nestjs/jwt';
 import { jwtConstants } from './constants';
 import { AuthGuard } from './auth.guard';
-import { SqsProducerModule } from 'src/sqs-producer/sqs-producer.module';
+import { SqsProducerModule } from 'src/aws/sqs-producer/sqs-producer.module';
 
 @Module({
   imports: [
