+++ node-server/src/aws/sqs-producer/constant.ts
