+++ node-server/src/users/users.module.ts
@@ -3,7 +3,7 @@ import { MongooseModule } from '@nestjs/mongoose';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { User, UserSchema } from './schemas/user.schema';
-import { SqsProducerModule } from 'src/sqs-producer/sqs-producer.module';
+import { SqsProducerModule } from 'src/aws/sqs-producer/sqs-producer.module';
 
 @Module({
   imports: [
